backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.11.18",
    "aiosignal>=1.3.2",
    "aiosqlite>=0.21.0",
    "alembic>=1.12.1",
    "annotated-types>=0.7.0",
    "anyio>=4.9.0",
    "attrs>=25.3.0",
    "bandit>=1.8.3",
    "bcrypt>=4.3.0",
    "beautifulsoup4>=4.13.4",
    "boto3>=1.38.43",
    "celery>=5.5.3",
    "certifi>=2025.1.31",
    "charset-normalizer>=3.4.2",
    "click>=8.1.8",
    "dataclasses-json>=0.6.7",
    "distro>=1.9.0",
    "dnspython>=2.7.0",
    "email-validator>=2.2.0",
    "fastapi-mail>=1.4.1",
    "fastapi>=0.104.0",
    "frozenlist>=1.6.0",
    "greenlet>=3.2.1",
    "h11>=0.14.0",
    "httpcore>=1.0.8",
    "httpx>=0.28.1",
    "idna>=3.10",
    "jiter>=0.9.0",
    "jsonpatch>=1.33",
    "jsonpointer>=3.0.0",
    "langchain-community>=0.2.19",
    "langchain-core>=0.1.53",
    "langchain-openai>=0.0.2",
    "langchain-text-splitters>=0.3.0",
    "langchain>=0.3.0.dev1",
    "langsmith>=0.1.147",
    "marshmallow>=3.26.1",
    "multidict!=6.5.0",
    "mypy-extensions>=1.1.0",
    "numpy>=1.26.4",
    "openai>=1.3.0",
    "orjson>=3.10.16",
    "packaging>=23.2",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.2.1",
    "pre-commit>=4.2.0",
    "prometheus-client>=0.22.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "propcache>=0.3.1",
    "psycopg2-binary>=2.9.9",
    "py-spy>=0.4.0",
    "pyasn1>=0.4.8",
    "pydantic-core>=2.33.1",
    "pydantic>=2.5.0",
    "pytest-xdist>=3.7.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0.2",
    "redis>=6.2.0",
    "regex>=2024.11.6",
    "requests-toolbelt>=1.0.0",
    "requests>=2.32.3",
    "rsa>=4.9.1",
    "ruff>=0.11.12",
    "six>=1.17.0",
    "slowapi>=0.1.9",
    "sniffio>=1.3.1",
    "sqlalchemy>=2.0.23",
    "starlette>=0.46.2",
    "tenacity>=8.5.0",
    "tiktoken>=0.9.0",
    "tqdm>=4.67.1",
    "typing-extensions>=4.13.2",
    "typing-inspect>=0.9.0",
    "typing-inspection>=0.4.0",
    "urllib3>=2.4.0",
    "uvicorn>=0.24.0",
    "yarl>=1.20.0",
    "pytest-asyncio>=1.0.0",
    "web3>=7.12.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

# Exclude specific directories and files
exclude = [
    ".venv",
    "__pycache__",
    "migrations",
    "alembic",
    "tests",
    "tarot_image_utils",
]

[tool.ruff.lint]
# Enable pycodestyle (E) and Pyflakes (F) codes by default
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "DJ", "PIE", "SIM", "TID", "PTH"]
ignore = [
"A003",
"B008",
"B904",
"B905",
"E501",
"UP007",
"N805",
"N806",
"N818"
]  # Various rules temporarily ignored
# Enable additional rules
extend-select = ["I"]  # isort
unfixable = ["B"]  # Avoid auto-fixing potentially dangerous fixes

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Use spaces for indentation
indent-style = "space"

[tool.bandit]
# Bandit security configuration
exclude_dirs = ["tests", "migrations", "alembic", ".venv"]
skips = ["B101", "B601", "B113", "B311"]  # Skip assert statements, shell injection (paramiko), request timeout, and pseudo-random generators

[tool.bandit.assert_used]
skips = ["**/tests/**", "**/test_*.py"]

[tool.setuptools]
packages = ["routers", "tarot_image_utils"]
